# vectorize `fold_even_odd`