--- a/fri/src/fold_even_odd.rs
+++ b/fri/src/fold_even_odd.rs
@@ -1,7 +1,8 @@
+use alloc::vec;
 use alloc::vec::Vec;
 
 use itertools::izip;
-use p3_field::{Powers, TwoAdicField};
+use p3_field::{PackedField, Powers, TwoAdicField};
 use p3_util::log2_strict_usize;
 use tracing::instrument;
 
@@ -28,6 +29,50 @@ pub fn fold_even_odd<F: TwoAdicField>(poly: &[F], beta: F) -> Vec<F> {
 
     let n = poly.len();
     debug_assert!(n > 1);
+
+    if n < F::Packing::WIDTH {
+        let mut res = vec![F::ZERO; n / 2];
+        let (first, second) = poly.split_at(n / 2);
+        fold_even_odd_packed::<F, F>(res.iter_mut(), first.iter(), second.iter(), n, beta);
+        res
+    } else {
+        let half_n = n / 2;
+        let nearest_mutliple_of_packing_width = (half_n + F::Packing::WIDTH - 1) / F::Packing::WIDTH;
+        let cutoff = (half_n / F::Packing::WIDTH) * F::Packing::WIDTH;
+
+        let mut res = vec![F::ZERO; nearest_mutliple_of_packing_width];
+        let res_packed = F::Packing::pack_slice_mut(&mut res);
+
+        let (first, second) = poly.split_at(n / 2);
+        let first_leftover =
+            F::Packing::from_fn(|i| if cutoff + i < first.len() { first[i] } else { F::ZERO });
+        let second_leftover = F::Packing::from_fn(|i| {
+            if cutoff + i < second.len() {
+                second[cutoff + i]
+            } else {
+                F::ZERO
+            }
+        });
+        let first = F::Packing::pack_slice(&first[..cutoff])
+            .iter()
+            .chain(core::iter::once(&first_leftover));
+        let second = F::Packing::pack_slice(&second[..cutoff])
+            .iter()
+            .chain(core::iter::once(&second_leftover));
+
+        fold_even_odd_packed::<F, F::Packing>(res_packed.iter_mut(), first, second, n, beta);
+        res.truncate(half_n);
+        res
+    }
+}
+
+fn fold_even_odd_packed<'a, F: TwoAdicField, P: PackedField<Scalar = F>>(
+    dst: impl Iterator<Item = &'a mut P>,
+    first: impl Iterator<Item = &'a P>,
+    second: impl Iterator<Item = &'a P>,
+    n: usize,
+    beta: F,
+) {
     let log_n = log2_strict_usize(n);
 
     let g_inv = F::two_adic_generator(log_n).inverse();
@@ -36,12 +81,15 @@ pub fn fold_even_odd<F: TwoAdicField>(poly: &[F], beta: F) -> Vec<F> {
 
     // beta/2 times successive powers of g_inv
     let powers = Powers {
-        base: g_inv,
-        current: half_beta,
+        base: P::from_fn(|_| g_inv),
+        current: P::from_fn(|_| half_beta),
     };
 
-    let (first, second) = poly.split_at(n / 2);
-    izip!(powers, first, second)
-        .map(|(power, a, b)| (one_half + power) * *a + (one_half - power) * *b)
-        .collect()
+    let one_half = P::from_fn(|_| one_half);
+    for (src, dst) in izip!(powers, first, second)
+        .map(|(power, &a, &b)| (one_half + power) * a + (one_half - power) * b)
+        .zip(dst)
+    {
+        *dst = src;
+    }
 }