Summary of changes:
- use `F::Packing` in `fold_even_odd`, padding to the nearest multiple of `F::Packing::WIDTH` and truncating the result if necessary.
- add `PackedPowers`, a version of `Powers` that's more or less a "chunked" version of `Powers`, where the chunks are packed field elements.
- add a very basic test for `fold_even_odd`


Resulted in a significant (~3.5-4x) speedup for`BabyBear`, and small changes for other fields (which don't yet have vectorized impls)

For instance, on my M1 MBP, I get the following results for `fold_even_odd` size 2^16:

| field               | before (mean) | after (mean) | change           |
|---------------------|---------------|--------------|------------------|
| `BabyBear`          | 193.02 µs     | 45.240 µs    | -76.552% (~4.2x) |
| `Goldilocks`        | 61.270 µs     | 65.994 µs    | +7.6771%         |
| `Mersenne31Complex` | 666.26 µs     | 631.55 µs    | -5.0013%         |